decode - handle unexpected end of stream
pull streams emit 'true' or an error on drain.
( see https://github.com/pull-stream/pull-stream#source-aka-readable )

I was getting a case where the source was unexpectedly ending, and 'true' was being passed up as the error. `decodeFromReader` seemed like the correct place to handle this case, as it is the end of pull streams and the start of a error-first-cb api.